fix: Store empty static String field as nil in TF state
The 'static' String fields (description/dns_name) when empty are
returned via the Netbox API as an empty String `""` but they cannot
be cleared by POST/PATCH-ing an emptry String or null value. They are
cleared by setting a single whitespace character `" "`, before this
change these empty fields were always stored in the TF state as an existent
resource property even though they were unset. This also resulted in
these whitespace values being sent to Netbox every time a resource was
created/updated.

This change stores unset fields as `nil` and sets the default values of
the description and dns_name property to `nil`. Empty String fields are
stored as nil when reading from Netbox and nil fields are POST/PATCH-ed
to Netbox as the single whitespace String `" "`. When unset the
properties will not be sent in the request body to Netbox and will not
show in the TF show command.